Remove archived Vercel (OAuth) app
Refactor Vercel (token auth) app & components
