+++ components/vercel_token_auth/actions/cancel-deployment/cancel-deployment.mjs
@@ -1,14 +1,35 @@
-/* eslint-disable pipedream/required-properties-name */
-/* eslint-disable pipedream/required-properties-description */
-/* eslint-disable pipedream/required-properties-version */
-/* eslint-disable pipedream/required-properties-type */
-import base from "../../../vercel/actions/cancel-deployment/cancel-deployment.mjs";
-import overrideApp from "../../common/override-app.mjs";
-
-overrideApp(base);
+import vercelTokenAuth from "../../vercel_token_auth.app.mjs";
 
 export default {
-  ...base,
   key: "vercel_token_auth-cancel-deployment",
-  version: "0.0.2",
+  name: "Cancel Deployment",
+  description: "Cancel a deployment which is currently building. [See the docs](https://vercel.com/docs/rest-api#endpoints/deployments/cancel-a-deployment)",
+  version: "0.0.3",
+  type: "action",
+  props: {
+    vercelTokenAuth,
+    deployment: {
+      propDefinition: [
+        vercelTokenAuth,
+        "deployment",
+        () => ({
+          state: "BUILDING",
+        }),
+      ],
+    },
+    team: {
+      propDefinition: [
+        vercelTokenAuth,
+        "team",
+      ],
+    },
+  },
+  async run({ $ }) {
+    const params = {
+      teamId: this.team,
+    };
+    const res = await this.vercelTokenAuth.cancelDeployment(this.deployment, params, $);
+    $.export("$summary", `Successfully canceled deployment ${this.deployment}`);
+    return res;
+  },
 };
