+++ components/vercel_token_auth/actions/list-deployments/list-deployments.mjs
@@ -1,14 +1,48 @@
-/* eslint-disable pipedream/required-properties-name */
-/* eslint-disable pipedream/required-properties-description */
-/* eslint-disable pipedream/required-properties-version */
-/* eslint-disable pipedream/required-properties-type */
-import base from "../../../vercel/actions/list-deployments/list-deployments.mjs";
-import overrideApp from "../../common/override-app.mjs";
-
-overrideApp(base);
+import vercelTokenAuth from "../../vercel_token_auth.app.mjs";
 
 export default {
-  ...base,
   key: "vercel_token_auth-list-deployments",
-  version: "0.0.2",
+  name: "List Deployments",
+  description: "List deployments under the account corresponding to the API token. [See the docs](https://vercel.com/docs/rest-api#endpoints/deployments/list-deployments)",
+  version: "0.0.3",
+  type: "action",
+  props: {
+    vercelTokenAuth,
+    project: {
+      propDefinition: [
+        vercelTokenAuth,
+        "project",
+      ],
+    },
+    team: {
+      propDefinition: [
+        vercelTokenAuth,
+        "team",
+      ],
+    },
+    state: {
+      propDefinition: [
+        vercelTokenAuth,
+        "state",
+      ],
+    },
+    max: {
+      propDefinition: [
+        vercelTokenAuth,
+        "max",
+      ],
+    },
+  },
+  async run({ $ }) {
+    const params = {
+      projectId: this.project,
+      state: this.state,
+      teamId: this.team,
+    };
+    const res = await this.vercelTokenAuth.listDeployments(params, this.max, $);
+    $.export("$summary", `Found ${res.length} deployment${res.length !== 1
+      ? "s"
+      : ""}`);
+    return res;
+  },
 };
