+++ components/vercel_token_auth/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/vercel_token_auth",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Pipedream Vercel (token-based auth) Components",
   "main": "vercel_token_auth.app.mjs",
   "keywords": [
@@ -11,5 +11,8 @@
   "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "@pipedream/platform": "^1.2.0"
   }
 }
